--- a/be/src/vec/exec/format/parquet/vparquet_reader.cpp
+++ b/be/src/vec/exec/format/parquet/vparquet_reader.cpp
@@ -46,30 +46,43 @@ ParquetReader::~ParquetReader() {
 
 void ParquetReader::_init_profile() {
     if (_profile != nullptr) {
+        static const char* const parquetProfile = "ParquetReader";
+        ADD_TIMER(_profile, parquetProfile);
+
         _parquet_profile.filtered_row_groups =
-                ADD_COUNTER(_profile, "ParquetFilteredGroups", TUnit::UNIT);
+                ADD_CHILD_COUNTER(_profile, "FilteredGroups", TUnit::UNIT, parquetProfile);
         _parquet_profile.to_read_row_groups =
-                ADD_COUNTER(_profile, "ParquetReadGroups", TUnit::UNIT);
+                ADD_CHILD_COUNTER(_profile, "ReadGroups", TUnit::UNIT, parquetProfile);
         _parquet_profile.filtered_group_rows =
-                ADD_COUNTER(_profile, "ParquetFilteredRowsByGroup", TUnit::UNIT);
+                ADD_CHILD_COUNTER(_profile, "FilteredRowsByGroup", TUnit::UNIT, parquetProfile);
         _parquet_profile.filtered_page_rows =
-                ADD_COUNTER(_profile, "ParquetFilteredRowsByPage", TUnit::UNIT);
+                ADD_CHILD_COUNTER(_profile, "FilteredRowsByPage", TUnit::UNIT, parquetProfile);
         _parquet_profile.filtered_bytes =
-                ADD_COUNTER(_profile, "ParquetFilteredBytes", TUnit::BYTES);
-        _parquet_profile.to_read_bytes = ADD_COUNTER(_profile, "ParquetReadBytes", TUnit::BYTES);
-        _parquet_profile.column_read_time = ADD_TIMER(_profile, "ParquetColumnReadTime");
-        _parquet_profile.parse_meta_time = ADD_TIMER(_profile, "ParquetParseMetaTime");
+                ADD_CHILD_COUNTER(_profile, "FilteredBytes", TUnit::BYTES, parquetProfile);
+        _parquet_profile.to_read_bytes =
+                ADD_CHILD_COUNTER(_profile, "ReadBytes", TUnit::BYTES, parquetProfile);
+        _parquet_profile.column_read_time =
+                ADD_CHILD_TIMER(_profile, "ColumnReadTime", parquetProfile);
+        _parquet_profile.parse_meta_time =
+                ADD_CHILD_TIMER(_profile, "ParseMetaTime", parquetProfile);
 
         _parquet_profile.file_read_time = ADD_TIMER(_profile, "FileReadTime");
         _parquet_profile.file_read_calls = ADD_COUNTER(_profile, "FileReadCalls", TUnit::UNIT);
         _parquet_profile.file_read_bytes = ADD_COUNTER(_profile, "FileReadBytes", TUnit::BYTES);
-        _parquet_profile.decompress_time = ADD_TIMER(_profile, "ParquetDecompressTime");
+        _parquet_profile.decompress_time =
+                ADD_CHILD_TIMER(_profile, "DecompressTime", parquetProfile);
         _parquet_profile.decompress_cnt =
-                ADD_COUNTER(_profile, "ParquetDecompressCount", TUnit::UNIT);
-        _parquet_profile.decode_header_time = ADD_TIMER(_profile, "ParquetDecodeHeaderTime");
-        _parquet_profile.decode_value_time = ADD_TIMER(_profile, "ParquetDecodeValueTime");
-        _parquet_profile.decode_dict_time = ADD_TIMER(_profile, "ParquetDecodeDictTime");
-        _parquet_profile.decode_level_time = ADD_TIMER(_profile, "ParquetDecodeLevelTime");
+                ADD_CHILD_COUNTER(_profile, "DecompressCount", TUnit::UNIT, parquetProfile);
+        _parquet_profile.decode_header_time =
+                ADD_CHILD_TIMER(_profile, "DecodeHeaderTime", parquetProfile);
+        _parquet_profile.decode_value_time =
+                ADD_CHILD_TIMER(_profile, "DecodeValueTime", parquetProfile);
+        _parquet_profile.decode_dict_time =
+                ADD_CHILD_TIMER(_profile, "DecodeDictTime", parquetProfile);
+        _parquet_profile.decode_level_time =
+                ADD_CHILD_TIMER(_profile, "DecodeLevelTime", parquetProfile);
+        _parquet_profile.decode_null_map_time =
+                ADD_CHILD_TIMER(_profile, "DecodeNullMapTime", parquetProfile);
     }
 }
 
@@ -97,6 +110,8 @@ void ParquetReader::close() {
             COUNTER_UPDATE(_parquet_profile.decode_dict_time, _column_statistics.decode_dict_time);
             COUNTER_UPDATE(_parquet_profile.decode_level_time,
                            _column_statistics.decode_level_time);
+            COUNTER_UPDATE(_parquet_profile.decode_null_map_time,
+                           _column_statistics.decode_null_map_time);
         }
         _closed = true;
     }